requirements.txt
matplotlib
numpy
panel<1.6
pyyaml
watchfiles
